requirements.txt
# --- Core Data Processing ---
pandas==2.2.1
pyarrow>=14.0.0
numpy<2.0.0

# --- Cloud Connectors ---
google-cloud-bigquery==3.19.0
databricks-sql-connector==3.1.0

# --- Auth & API Handling ---
requests==2.31.0
python-dotenv==1.0.1
google-auth==2.28.1

# --- Performance & Compatibility ---
thrift==0.16.0
# --- SSL & Certificate Handling ---
pip-system-certs==4.0
certifi>=2024.2.2